+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Breaking: Bump minimum required C++ standard from C++17 to C++20. (#55)
 - Minor: Added support for (de-)serializing from/to [std::variant](https://en.cppreference.com/w/cpp/utility/variant). (#54)
+- Minor: Added support for (de-)serializing from/to [std::optional](https://en.cppreference.com/cpp/utility/optional). (#59)
 
 ## v0.2.1